moved all fragments to inside MainActivity
