+++ .idea/modules.xml
@@ -2,7 +2,9 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/AndroidCarbonTracker.iml" filepath="$PROJECT_DIR$/AndroidCarbonTracker.iml" />
       <module fileurl="file://$PROJECT_DIR$/CarbonTracker.iml" filepath="$PROJECT_DIR$/CarbonTracker.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
 </project>
